chore: mock data 위치 변경
